Adds ability to list namespaces in JSON format (#6295)
* Adds ability to list namespaces in JSON format

Signed-off-by: Dharmit Shah <shahdharmit@gmail.com>

* Tests without Expect

Using `Expect` directly or through something that uses it under the hood
is a problem when using `Eventually`. In normal case, a failure in
matching test condition behaves like `continue` and `Evnetually` keeps
trying again till timeout is reached. But `Expect` changes that
behaviour to be like `break` and doesn't keep testing repeatedly as we
would like it to.

Signed-off-by: Dharmit Shah <shahdharmit@gmail.com>

* UI fixes when there's no project on cluster

Signed-off-by: Dharmit Shah <shahdharmit@gmail.com>

* Adds an import to fix unit tests

Signed-off-by: Dharmit Shah <shahdharmit@gmail.com>

* More JSON specific checks

Signed-off-by: Dharmit Shah <shahdharmit@gmail.com>

* Consistent JSON output for namespaces

JSON output for components, bindings and services when odo can't find
any of these resources is simply a `{}`. This commit makes output of
`odo list namespaces -o json` consistent with that of these commands.

Signed-off-by: Dharmit Shah <shahdharmit@gmail.com>

* Add doc for namespaces JSON output

Signed-off-by: Dharmit Shah <shahdharmit@gmail.com>

* Test if some other namespace is marked as active: false

Signed-off-by: Dharmit Shah <shahdharmit@gmail.com>

Signed-off-by: Dharmit Shah <shahdharmit@gmail.com>