Implement `odo delete component --files` to delete files generated by `odo` (#6255)
* Add utility functions allowing to report files generated by odo

The end goal is for 'odo delete component --files' to gather a list of
files that were initially generated by odo, so as to delete only those.
This way, we are less likely to delete the wrong files.

This list is collected in a '.odo/generated' file.

It will then be up to odo commands to call those methods
accordingly. For example, this is called:
- by odo init, odo dev, or odo deploy, when generating a new devfile.yaml
- by odo dev, when there is no .gitignore file and odo generated one

Note that it is quite impossible to cover all scenarios; for example,
if odo generates a new devfile.yaml file, but the user deletes it
manually and recreates it, 'odo delete component --files'
will still list it as a possible candidate for deletion
(because it would be listed in .odo/generated).

* Report files generated by commands like init, dev or deploy

* Implement 'odo delete component --files'

* Add integration tests for 'odo delete component --files'

* Document the `--files` flag

Co-authored-by: Philippe Martin <contact@elol.fr>

* Do not delete the .gitignore file

It is more likely to be modified by the user afterward (for another
usage).

Co-authored-by: Philippe Martin <phmartin@redhat.com>

* fixup! Document the `--files` flag

* fixup! Implement 'odo delete component --files'

* Apply review suggestions on integration tests

Co-authored-by: Parthvi Vala <pvala@redhat.com>

* fixup! fixup! Document the `--files` flag

Co-authored-by: Philippe Martin <contact@elol.fr>
Co-authored-by: Philippe Martin <phmartin@redhat.com>
Co-authored-by: Parthvi Vala <pvala@redhat.com>