Tell users of the QuickStart Guide to select "Maven" on the Spring Initializr [1] (#6469)
It looks like the default "Project" on [1] is now "Gradle - Groovy" (it used to be Maven),
but there is no Devfile usable with a SpringBoot Gradle project.

Without this, "odo init" prompts to select the "java-openliberty-gradle" Devfile,
which won't work because it executes tasks specific to OpenLiberty.

And even if we manually select the "java-springboot" Devfile,
it won't work out of the box because it requires Maven and there is
no "pom.xml" in the project generated by the Spring Initializr.

[1] https://start.spring.io/