Fix potential issues with E2E tests (#6543)
* Make 'helper.Component#Exec' accept an additional parameter that allows checking (or not) the remote command execution

* In E2E tests, wait until the app is completely ready before sending requests to the local forwarded port

Otherwise, sending a request too early causes
"connection refused" errors, which causes port-forwarding to be restarted
(with new local ports because "odo dev" is running with "--random-ports").
As a first start, here we are making sure that the application port is available in the container,
prior to sending requests to the local forwarded port. This should reduce the likeliness
of port-forwarding restarts.

In the future, we might want to make the test detect the new local ports and use those instead.
We could even make sure to make "odo dev" reuse the previous random ports it had, provided
they are available again.

* Increase timeout and polling interval when waiting for port-forward restart

* Add a '/ping' endpoint to the sample Go app used in E2E

This allows checking the application running in the container.

* Do not stop the DevSession process before removing the Binding from the Devfile

https://github.com/redhat-developer/odo/issues/6101 has been fixed.

* fixup! Make 'helper.Component#Exec' accept an additional parameter that allows checking (or not) the remote command execution

* fixup! In E2E tests, wait until the app is completely ready before sending requests to the local forwarded port

* Create reusable 'helper.WaitAppReadyInContainer' helper function, as suggested in review

Co-authored-by: Philippe Martin <phmartin@redhat.com>

---------

Co-authored-by: Philippe Martin <phmartin@redhat.com>