Wait until expected ports are opened in the container before starting port-forwarding (#6701)
* Add function in 'port' package allowing to check whether given ports are actually opened and in LISTEN state in specified containers

* Wait on Kubernetes until the application is ready by checking the ports to forward to are actually opened (or a timeout expires)

This allows to display port-forwarding messages when the application is ready to accept requests.
Otherwise, if the application takes time to listen on the port,
and we try to reach the local forwarded port, port forwarding will
be restarted.
If we are using --random-ports, new random ports will be generated.

Known issue: with the backo-go exponential backoff implementation,
it seems hitting Ctrl-C does not stop waiting =>
we need to wait until the timeout is reached or the backoff is done.

* Wait on Podman until the application is ready by checking the ports to forward to are actually opened (or a timeout expires)

This allows to display port-forwarding messages when the application is ready to accept requests.
Otherwise, if the application takes time to listen on the port,
and we try to reach the local forwarded port, port forwarding will
be restarted.
If we are using --random-ports, new random ports will be generated.

Known issue: with the backo-go exponential backoff implementation,
it seems hitting Ctrl-C does not stop waiting =>
we need to wait until the timeout is reached or the backoff is done.

* Display warnings instead if the ports are not opened

While iterating on the application, it might feel
weird to stop the Dev Session immediately.
Plus, a lot of integration tests are not passing because
of source code not strictly matching the ports declared in the
(too many) Devfiles.

* Fix integration test checking the behavior when simulating an app taking long to start

* Fix output expected for documentation tests

* Add hint to run 'odo logs --follow [--platform]' to help understand why app is not listening on the expected ports

Co-authored-by: Parthvi Vala <pvala@redhat.com>

* Collect all unreachable ports along with the containers and return the list to the users

Co-authored-by: Parthvi Vala <pvala@redhat.com>

* Apply suggestions from code review

Co-authored-by: Parthvi Vala <pvala@redhat.com>

* fixup! Add hint to run 'odo logs --follow [--platform]' to help understand why app is not listening on the expected ports

* fixup! Add hint to run 'odo logs --follow [--platform]' to help understand why app is not listening on the expected ports

---------

Co-authored-by: Parthvi Vala <pvala@redhat.com>