Fix flaky tests using custom port mapping when running specs in parallel (#6763)
* Fix flaky test using custom port mapping when running specs in parallel

Using a global counter currently does not work
because Ginkgo runs parallel specs by isolating them in
completely different processes [1].
This relies on the parallel process index returned by Ginkgo
to get a value that can be used as starting port.
Callers can then safely increment this value if needed in the
specs to get a different custom value.

[1] https://onsi.github.io/ginkgo/#mental-model-how-ginkgo-runs-parallel-specs

* Apply suggestions from code review

Co-authored-by: Parthvi Vala <pvala@redhat.com>

* fixup! Apply suggestions from code review

---------

Co-authored-by: Parthvi Vala <pvala@redhat.com>