Replace deprecated `--slow-spec-threshold` Ginkgo flag with `--poll-progress-after` (#6834)
* Replace deprecated '--slow-spec-threshold' Ginkgo flag with '--poll-progress-after'

Per [1], this instructs Ginkgo to automatically emit a progress report
whenever a node takes too long to complete.
The Progress Report includes information about which node is currently running
and the exact line of code that it is currently executing,
along with any relevant goroutines that were launched by the spec.

[1] https://onsi.github.io/ginkgo/#getting-visibility-into-long-running-specs

* Also set '--poll-progress-interval' flag so as to get progress reports periodically