Clean up OWNERS files
This makes sure the list is accurate, to avoid requesting PR reviews
from people no longer maintaining the project.
Ideally, I would put a GitHub team,
but this does not seem to be supported [1].
It can be done via an additional OWNERS_ALIASES file [2],
but this is yet to be tested.

[1] https://www.kubernetes.dev/docs/guide/owners/
[2] https://www.kubernetes.dev/docs/guide/owners/#owners_aliases