Make tests start an HTTP server service Devfile artifacts
This allows us to remove the dependency on a Devfile Registry server.
We no longer want to have a long-running cluster
to which a Devfile registry is deployed.
And relying on the public staging registry
makes those tests sometimes fail, due
to the frequency of updates and changes that break
our tests.
Now, an HTTP Server instance is started
on a random ports and serves the OCI artifacts for
the stacks that are useful for our tests.

Note that this will now be the default behavior,
unless the DEVFILE_REGISTRY env var is explicitly set.

This allows us to also update the stacks at our own pace.
The process for updating is as follows:
1- update the stacks folder under tests/helper/registry_server/testdata/registry
2- run 'generate-test-registry-build' to regenerate the Devfile Index
and OCI archives for stack extra-files
3- Commit and push the changes